Directional lights inplemented. Broken
The light seems to spin around the cube a bit faster than the cube is
spinning and not staying in place.
HLSL light structs now in hlsl header file and just included in the
shader to save time rewriting the code.
Only using one constant buffer, it's inefficent but i still cant figure
out how to use multiple.