Multiple constant buffers and working specular
Now rendering both the sharp edged cube and the soft cube next to each
other on the screen.
Now need to do point lights and spot lights.

The issue with the multiple constant buffers came down to manually
specifying the register in the shader and cbuffers need to have a size
that's a multiple of 16.